rebase

+++ sdk/core/Azure.Core/tests/CoreResponseClassifierTests.cs
@@ -12,9 +12,9 @@ namespace Azure.Core.Tests
         public void ClassifiesSingleCodeAsNonError()
         {
             // test classifiers for each of the status codes
-            for (int nonError = 100; nonError <= 599; nonError++)
+            for (ushort nonError = 100; nonError <= 599; nonError++)
             {
-                CoreResponseClassifier classifier = new CoreResponseClassifier(new int[] { nonError });
+                CoreResponseClassifier classifier = new CoreResponseClassifier(new ushort[] { nonError });
                 HttpMessage message = new HttpMessage(new MockRequest(), classifier);
 
                 // test all the status codes against the classifier
@@ -43,7 +43,7 @@ namespace Azure.Core.Tests
         [TestCase(502, true)]
         public void ClassifiesMultipleCodesAsNonErrors(int code, bool isError)
         {
-            CoreResponseClassifier classifier = new CoreResponseClassifier(stackalloc int[] { 200, 404 });
+            CoreResponseClassifier classifier = new CoreResponseClassifier(stackalloc ushort[] { 200, 404 });
 
             HttpMessage message = new HttpMessage(new MockRequest(), classifier);
